Progress in Map tab. Create event changes
An event is now created from Map tab by long press in any location from the map. A (now rounded-corner) dialog box opens. The location is added now. If user not agrees to location the app will not start (the code for permission is in MainActivity), this behaviour can be changed later.