Small Vibration and bugfix
Small vibration when opening create event menu, style update for night mode (Not available in the app yet).
When an event was deleted or it's far away from user's location, it still was in the joined events list, but the information for that event was unavailable. Now you can unjoin without crashes.